Move ChromaticityMonitor from diagnostics to tuning tool.
