+++ examples/SOLEIL_examples/tuning_tools.yaml
@@ -16,7 +16,7 @@
   quad_array_name: QCORR
   betatron_tune_name: BETATRON_TUNE
   quad_delta: 1e-3
-- type: pyaml.diagnostics.chromaticity_monitor
+- type: pyaml.tuning_tools.chromaticity_monitor
   name: DEFAULT_CHROMATICITY_MEASUREMENT
   betatron_tune_name: BETATRON_TUNE
   rf_plant_name: RF
